backend-ci/requirements.txt
# Django Core ...
asgiref==3.11.0
attrs==25.4.0
Django==4.2
sqlparse==0.5.5
tzdata==2025.3

# Django Extensions ...
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1

# API Documentation
drf-spectacular==0.29.0
inflection==0.5.1
uritemplate==4.2.0

# JSON Schema
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
referencing==0.37.0
rpds-py==0.30.0

# JWT
PyJWT==2.10.1
PyYAML==6.0.3

# Type Hints
typing_extensions==4.15.0

# Production Server
gunicorn==21.2.0

# PostgreSQL Database
psycopg2-binary==2.9.9

# Environment Variables
python-decouple==3.8
